Move STABS to AST conversion code into its own file
